requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
djangorestframework==3.12.4
idna==3.3
praw==7.4.0
prawcore==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.1
